+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Experimental support for listing and deleting all resources with `all list` and `all purge` commands. This initial version supports networks, network-peerings, routers, databases, and object-storages.
+- Add `--wait` flag to `object-storage delete` command.
+- Add `--disable-termination-protection` and `--wait` flags to `database delete` command.
+
 ### Fixed
 
 - Kubernetes node-group subcommand not working
@@ -14,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 
 - Deprecated `upctl kubernetes nodegroup` command, use `upctl kubernetes node-group` instead.
+
 ## [3.17.0] - 2025-03-14
 
 ### Added