+++ CHANGELOG.md
@@ -543,7 +543,8 @@ Initial public beta release :tada:
 ### Added
 - Current feature set added! First internal release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.17.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.18.0...HEAD
+[3.17.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.17.0...v3.18.0
 [3.17.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.16.1...v3.17.0
 [3.16.1]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.16.0...v3.16.1
 [3.16.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v3.15.0...v3.16.0